requirements_dev.txt
pip==8.1.2
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==2.6.0
tox==2.3.1
coverage==4.1
Sphinx==1.4.4
cryptography==1.4
PyYAML==3.11
requests>=2.11.1
ansicolors>=1.0.2
PyYAML>=3.12
